Update go support plugin, go path addition
